vimrc: turn off 'display' motions
If a line wraps, hitting 'j', for example, takes you to the next line,
not the next display line. I had remapped the up/down keys to prefix with
the 'g' (e.g. gj and gk) but I prefer the default behaviour, and I can
remember to use the 'g' prefix when I need (and I can use it in more
cases, e.g. g$ to move to the end of the display line).