fvwm2rc: shift+pgup/pgdn in scroll-bar
This is odd, and a new behaviour since upgrading to openSuSE 12.3.

Previously I had set the key binding Alt+Tab to move my mouse pointer to the
lower right corner (-1 -1) of the window. If I then used Shift+PgUp or
Shift+PgDn to scroll everything was fine. Now, however, if the window's bottom
is along the bottom of the screen, the pointer is not placed at -1 -1 but
rather is placed so that the bottom of the arrow is at -1 -1 instead of the
hot-point.

This causes the pointer to be over the scroll bar of my xterm (which I have
placed on the right side of my window). When the pointer is in this position
(i.e. over the scroll bar) and I press Shift+PgUp or Shift+PgDn the entire
window disappears! Note, this only happens if the window is aligned with the
bottom of the screen; windows floating around have the pointer hot-spot placed
at -1 -1, so everything works.

So now I'm forcing the pointer to be placed at -1 -1 but then moved 35 pixels
to the right so that it is never placed on top of the scroll bar.