git: add aliases
These two aliases help identify information about a branch and the point at
which the branch branched off from whatever it is the branch branches off
from.

In one case it finds the branch's first commit after the branch-off point.
In the second case it finds the last common commit between a branch and its
branch-off point.

See http://stackoverflow.com/questions/1527234/finding-a-branch-point-with-git

Signed-off-by: Trevor Woerner <twoerner@gmail.com>