volequ.sh: not just ogg
Update this script (volequ.sh) so that it is not ogg-specific; make it
flexible enough to be able to process sound files of any type (e.g. mp3).

All that is really required is to not assume '.ogg' as the input filename
extension and simply find the extension by examining the input file.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>