bashrc: refactor the root vs user delimiter
Instead of maintaining 2 separate arrays of PS1 elements that only differ
in the colour of the delimiter ($_regdelim vs $_rootdelim), move the "root
vs user" logic higher and create only 1 delimiter that is one colour or the
other. This allows us to have just one array of PS1 elements that simply
contains $_delim.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>