bashrc: fix multiline calculation
If/when multiline commands are entered by the user, the code has to be smart
enough to move the cursor up a one or more lines before overwriting the dashes
with the current date/time stamp. The original code wasn't working anywhere
near correctly. As such, multiline commands would get part of their command
overwritten by the date/time stamp.

This solution was inspired by:
	https://superuser.com/questions/1497179/how-to-get-the-length-of-a-prompt
which (unfortunately?) depends on perl being available.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>