imgrot.sh: reset angle between images
Reset the "rotate" angle at the start of the loop so that each image is
processed for their angle separately (instead of setting the angle once
from the first pass through the loop). If an angle is specified on the
cmdline, it needs to be stored separately from the "rotate" variable.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>