mutt: fix patch-saving macro
The script that was being run to save tagged messages (mutt-savembox)
was not saving the messages in anything even closely resembling the mbox
format. Assuming we want a macro to save an email into a patch format,
rename the macro accordingly. Also, update the macro to use "formail" to
convert the tagged mail to a patch format.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>